Deploy/requirements.txt
streamlit
sklearn
pandas
tensorflow-cpu
numpy
plotly.express
